+++ src/components/Forms/Annocement/types.ts
@@ -1,0 +1,5 @@
+interface RouteParams {
+  addId: string;
+}
+
+export default RouteParams;
