+++ src/views/CreateAnnoucements/index.tsx
@@ -1,5 +1,5 @@
 import { Container } from '@material-ui/core';
-import AnnoucementForm from '../../components/Forms/annoucement.form';
+import AnnoucementForm from '../../components/Forms/Annocement';
 import ViewTitle from '../../components/ViewTitle';
 import useStyles from './styles';
